fix(new object): requested changes in the PR
